+++ src/page/SignUp/SignStep6.tsx
@@ -2,7 +2,7 @@ import { StepProps } from '@/type/SignUp';
 import { ReactComponent as IconArrowLeft } from '@/assets/icons/IconArrowLeft.svg';
 import { styled } from 'styled-components';
 import { Button } from '@/components/common/Button/Button';
-import { SearchInput } from '@/page/SignUp/SearchInput';
+import { SearchInput } from '@/components/common/SignUp/SearchInput';
 
 import { useState } from 'react';
 import { PlainInputBox } from '@/components/common/InputBox/PlainInputBox';
